# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for employee registration form with Odoo integration
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Employee {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  cpf?: string | null;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  jobPosition?: string | null;
  /** @nullable */
  hireDate?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  odooId?: number | null;
  createdAt?: string;
}

export type EmployeeInputGender = typeof EmployeeInputGender[keyof typeof EmployeeInputGender];


export const EmployeeInputGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface EmployeeInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  mobile?: string;
  cpf?: string;
  birthDate?: string;
  gender?: EmployeeInputGender;
  department?: string;
  jobTitle?: string;
  jobPosition?: string;
  hireDate?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
}

export interface Department {
  id: number;
  name: string;
}

export interface JobPosition {
  id: number;
  name: string;
}

export interface OdooStatus {
  connected: boolean;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  message?: string | null;
}