# lib/api-zod/src/generated/types/department.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for employee registration form with Odoo integration
 * OpenAPI spec version: 0.1.0
 */

export interface Department {
  id: number;
  name: string;
}