# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for employee registration form with Odoo integration
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  cpf?: string | null;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  jobPosition?: string | null;
  /** @nullable */
  hireDate?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  odooId?: number | null;
  createdAt?: string;
}