# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for employee registration form with Odoo integration
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputGender } from './employeeInputGender';

export interface EmployeeInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  mobile?: string;
  cpf?: string;
  birthDate?: string;
  gender?: EmployeeInputGender;
  department?: string;
  jobTitle?: string;
  jobPosition?: string;
  hireDate?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
}