# lib/api-zod/src/generated/types/odooStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for employee registration form with Odoo integration
 * OpenAPI spec version: 0.1.0
 */

export interface OdooStatus {
  connected: boolean;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  message?: string | null;
}